tooling/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isms-tooling"
version = "0.1.0"
description = "Tooling for the ISMS repository: schemas, validators, collectors, signers, packagers, renderer."
readme = "../README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [
    { name = "isms contributors" },
]
keywords = ["isms", "iso27001", "nis2", "nisg", "compliance", "governance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "jsonschema>=4.21",
    "ruamel.yaml>=0.18",
    "python-dateutil>=2.8",
    "requests>=2.31",
    "lxml>=5.0",
    "cryptography>=42.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.urls]
Repository = "https://github.com/OWNER/isms"

[tool.setuptools]
packages = []

[tool.setuptools.package-dir]
"" = "."

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
